Require CERN authentication
As of mai 2019, the run registry API is no longer accessible without logging in. Therefore the cernrequests package has to be used to retrieve sso cookies.
