+++ runreg/__init__.py
@@ -12,4 +12,4 @@
 
 from .client import get
 
-__version__ = "0.2.1"
+__version__ = "0.3.0"
