+++ runreg/client.py
@@ -11,8 +11,12 @@
 # or submit itself to any jurisdiction.
 
 import json
+
+import cernrequests
 import requests
 from runreg.utils import create_filter, flatten_runs, convert_lookup_fields
+
+base_url = "https://cmsrunregistry.web.cern.ch"
 
 url_format = (
     "https://cmsrunregistry.web.cern.ch/api/datasets/{workspace}/editable/{page}/"
@@ -27,14 +31,17 @@ def _create_payload(page_size=PAGE_SIZE, **kwargs):
     )
 
 
-def _get_page(page, workspace, **kwargs):
+def _get_page(page, workspace, cookies, **kwargs):
     url = url_format.format(workspace=workspace.lower(), page=page)
     headers = {"Content-type": "application/json"}
     payload = _create_payload(**(convert_lookup_fields(**kwargs)))
-    return requests.post(url, headers=headers, data=payload).json()
+    response = requests.post(url, headers=headers, data=payload, cookies=cookies)
+    return response.json()
 
 
-def get(flat=False, workspace="tracker", **kwargs):
+def get(flat=False, workspace="tracker", cookies=None, **kwargs):
+    if cookies is None:
+        kwargs["cookies"] = cernrequests.get_sso_cookies(base_url)
     initial_response = _get_page(0, workspace, **kwargs)
 
     if "err" in initial_response:
